*: RTFILE becomes a pointer, RTFileOpen++ expands it's flags paramter from uint32_t to uint64_t.
svn:sync-xref-src-repo-rev: r72442

+++ src/VBox/Runtime/common/misc/RTFileOpenF.cpp
@@ -1,4 +1,4 @@
-/* $Id: RTFileOpenF.cpp 28800 2010-04-27 08:22:32Z noreply@oracle.com $ */
+/* $Id: RTFileOpenF.cpp 37596 2011-06-22 19:30:06Z knut.osmundsen@oracle.com $ */
 /** @file
  * IPRT - RTFileOpenF.
  */
@@ -32,7 +32,7 @@
 #include "internal/iprt.h"
 
 
-RTR3DECL(int) RTFileOpenF(PRTFILE pFile, uint32_t fOpen, const char *pszFilenameFmt, ...)
+RTR3DECL(int) RTFileOpenF(PRTFILE pFile, uint64_t fOpen, const char *pszFilenameFmt, ...)
 {
     va_list va;
     va_start(va, pszFilenameFmt);
